another trial of the fibonacci

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,15 +8,15 @@
 int main(void)
 {
 	int count;
-	unsigned int fib1 = 1, fib2 = 2, fib_next;
+	unsigned long long fib1 = 1, fib2 = 2, fib_next;
 
-	printf("%u, %u", fib1, fib2);
+	printf("%llu, %llu", fib1, fib2);
 
 	for (count = 3; count <= 98; count++)
 	{
 		fib_next = fib1 + fib2;
 
-		printf(", %u", fib_next);
+		printf(", %llu", fib_next);
 
 		fib1 = fib2;
 		fib2 = fib_next;
